x/auction/keeper/msg_server_object_test.go
package keeper_test

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"

	"github.com/deepstack/auction/x/auction/types"
)

func TestObjectMsgServerCreate(t *testing.T) {
	srv, ctx := setupMsgServer(t)
	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreateObject(ctx, &types.MsgCreateObject{Creator: creator})
		require.NoError(t, err)
		require.Equal(t, i, int(resp.Id))
	}
}

func TestObjectMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdateObject
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateObject{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateObject{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateObject{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)
			_, err := srv.CreateObject(ctx, &types.MsgCreateObject{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdateObject(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestObjectMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeleteObject
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteObject{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteObject{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeleteObject{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)

			_, err := srv.CreateObject(ctx, &types.MsgCreateObject{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteObject(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
